Executable jar of Coronavirus tracker application
This application reads csv that's updated daily on github, parses it into string and then into objects. These objects are used to perform certain operations and display the data on a web page (using thymeleaf). Since the data is updated each day, the csv is read on a daily basis.